refactor: move access check functions to utils and update references
