+++ backend/versions/v1/controllers/lab/getLab.go
@@ -1,9 +1,9 @@
 package lab
 
 import (
+	"fmt"
 	"net/http"
 	"strconv"
-	"fmt"
 
 	"github.com/gin-gonic/gin"
 )
@@ -31,7 +31,7 @@ func (lc *LabController) GetLabByIDHandler(c *gin.Context) {
 		return
 	}
 
-	isEnrolledStudent, err := lc.Service.Model.CanStudentAccessLab(labIdInt, claims.UserID)
+	isEnrolledStudent, err := lc.Service.Utils.CanStudentAccessLab(labIdInt, claims.UserID)
 	isClassInstructor := lc.Service.Utils.IsUserTeacherAdminOrAssistantByLabID(labIdInt, claims.UserID)
 
 	if err != nil {
